Small changes to get a decent demo appearance when projecting from my laptop.
* index.html: shortened demo button names so they fit on a single
line.
* scripts/demo.js: added updateProcessState() to
updatePatientRecord(), so process state gets updated every time
the patient record does.  The functionality in demo.php should
probably be moved to peos.c so the process state is alway obtained
every time the server is contacted.
* scripts/peos.js: changed widths of panels to try to keep more of
the process outline in the pane.
* css/carepathways.css: changed all border colors to default (black).
* Makefile: all permissions now add write access to 'other' so the
scripts can create and update files.  This seems to be the only
way to easily and safely allow installation in public_html.

It's still not ideal due to the oddness with Pure grids overflowing,
but it's moving closer.  Probably need to nest divs inside the Pure
grid divs to take the styling.